requirements.txt
streamlit
numpy
torch
torchvision
Pillow>=7.1.0
scikit-image
matplotlib
click